refactor login logic with server code to now exclusively use session and passport, no local storage or JWT
