+++ src/main.js
@@ -3,7 +3,19 @@ import { createPinia } from "pinia";
 import "./style.css";
 import router from "./router/router.js";
 import App from "./App.vue";
+import { useAuthStore } from "./store/authStore.js";
 
-const pinia = createPinia();
+async function initialize() {
+  const app = createApp(App);
+  const pinia = createPinia();
 
-createApp(App).use(router).use(pinia).mount("#app");
+  app.use(pinia);
+
+  const authStore = useAuthStore();
+  await authStore.checkAuth();
+
+  app.use(router);
+  app.mount("#app");
+}
+
+initialize();
